Mbphotomosaic and mbgetphotocorrection
Mbphotomosaic: Fixed problem specifying the image correction table file in
an imagelist structure.

Mbphotomosaic and mbgetphotocorrection: Altered the image correction table file
generated by mbgetphotocorrection and read by mbphotomosaic to include
correction tables for color as well as intensity. Since we are using the YCbCr
color space this means generating tables of average chroma (color) difference
in red and blue as well as a 3D table of average luma (intensity).
If RGB and YCbCr values have ranges 0-255, then:
  Y  = 0.299R + 0.587G + 0.114B (Luminance, intensity, B/W signal)
  Cb = 0.564(B - Y) + 128 (blue color difference)
  Cr = 0.713(R - Y) + 128 (red color difference)
and
  B = (Y + 1.773 * (Cb - 128));
  G = (Y - 0.714 * (Cr - 128) - 0.344 * (Cb - 128));
  R = (Y + 1.403 * (Cr - 128));
If the specified color correction file includes the Cr and Cb tables, then the
--correct-color command will cause mbphotomosaic to use those tables to correct
for color. Each source image pixel being corrected has a lateral location within
the source image and a standoff (altitude) relative to the seafloor; these 3D
positions correspond to locations in the 3D correction tables. As with the
intensity, the Cr and Cb corrections are a multiplicative factor calculated to
map the the value in the relevant location in the table to a reference value.
The reference value may be the average central values in the correction tables
or a value specified by the user with the --reference-intensity=intensity and
--reference-crcb=cr/cb commands.