Mbphotomosaic and mbm_makeimagelist
Mbphotomosaic: Augmented available commands that can be specified in imagelist
files with --priority-weight and section-length-max. The --priority-weight allows
one to up-weight or down-weight sets of images in a photomosaic. The --section-length-max
command allows one to exclude groups of pixels that are projected into too large
of an extent when projected onto the seafloor topography model.

Mbm_makeimagelist: Augmented to comment out the first five image pairs in each
image survey directory, as the first three or four images are usually quite
dark.