Incremental changes to build system and MBnavadjust
Fixes to the Autotools build system allowing MB-System to be built in full on
Arm architecture Mac computers running the Monterey OS and fixing the build of
the TRN tools on Linux.

MBnavadjust: Incremental changes towards handling global ties relative to a
reference grid - defined a single section naverr mode but don't have it
working yet.