Mbtrnpp
Mbtrnpp: Added option --auv-sentry-em2040 to configuration file and command line
options in order to enable use of the special pressure depth encoding in EM2040
data collected on AUV Sentry. Previously this option was hard coded in mbtrnpp.c.
By default this option is disabled, so in order for mbtrnpp to work with Sentry
EM2040 multibeam either in realtime or playback modes this option must now be
specified.