Many fixes: mbnavadjust, mbareaclean, mbbackangle, mbprocess
Mbnavadjust: Added ability to import a reference bathymetry grid that can be
used to define global ties. Also modified so that when the main display list mode
limits what is shown to "with selected survey" or other modifiers, these constraints
also limit what navigation is shown in the project visualization. So, if one selects
a survey and then sets the list mode to show only the selected survey, then the
visualization will also only display the navigation from that survey, the crossing
ties or global ties from that survey, and only allow picking on that survey's
navigation.

Mbareaclean: Fixed issue in which mbareaclean does not apply filters for files
without pre-existing *.esf files.

Mbbackangle: Plots of the backscatter vs grazing angle distribution are no longer
histogram equalized.

Mbprocess: Modified the threaded processing function process_file() so that each
instance carries a thread id that can be checked during debugging.