Improved processing of Imagenex DeltaT multibeam data
Extended support of format MBF_IMAGE83P 191 to version 1.10 of the vendor format specification. This includes the adddition of beam intensity data used for per beam amplitude in MB-System.

Augmented the extended format MBF_IMAGEMBA 192 to also support version 1.10 and to carry the full calculated bathymetry and amplitude and to hold sonar depth values in the context of surveys from submerged platforms.

The program mbpreprocess now works fully with DeltaT data, including translating the logged format 191 to the extended format 192 for processing.

Mbpreprocess now has a kluge option to handle a situation where sonar depth values are recorded in the heave parameter of logged format 191 data. The option --kluge-sonardepth-from-heave moves the value from heave to sonardepth, leaving the heave as zero.