Fixed errors in last commit plus partial fix to Hysweep i/o module
Fixed some problems with the compiler warning fixes of the last commit.

Format 201 (MBF_HYSWEEP1): Fixed reading of Hysweep multibeam in which beam angles are
stored in spherical coordinates but are not corrected for roll and pitch. Writing these
data is still broken, and preprocessing to handle time latency issues with the attitude
data is not yet functional.