Prep for 5.7.9beta64, Major bug fix to reading GMT grids
Reading GMT grids: Fixed a very significant bug that caused GMT grids to be read into
memory incorrectly by mb_read_gmt_grd() in src/mbaux/mb_readwritegrd.c. The problem was
that the GMT default padding of grids by 2 grid cells around all edges was applied
incorrectly, with a result that the topography (or other data) in the grid was shifted
two grid cell widths north. The creation of grids by mbgrid or mbmosaic was correct;
the problem came when the grids were read in by mbgrdviz, mbanglecorrect, mbprocess,
and mbnavadjust.