Fix support for formats MBF_SBSIOSWB and MBF_EM710RAW
Format 16 (MBF_SBSIOSWB): Fixed bug that became evident when processing old Scripps
SeaBeam Classic data in format 16 on a MacOs Sonoma machine. This turned out to be
a poorly formed preprocessor macro for rounding floating point values in code dating
from 1992.

Formats 58 (MBF_EM710RAW) and 59 (MBF_EM710MBA):  Fixed bug in which some bathymetry
edits were applied to the wrong pings. Third generation Kongsberg multibeams produce
two cross profiles with each ping cycle, but represent these as two separate pings
with the same ping time. MB-System distinguishes between pings using the timestamps
rather than ping numbers (because not all sonars produce ping numbers), so pings
with the same timestamp confuse the beam flag handling code. The solution is to add
a small amount of time to the timestamp of the second ping in each pair. The bug was
that the amount of time added by mbpreprocess was 33 microseconds, but the datagrams
in Kongsberg *.all files specify timestamps only to the millisecond level, and so
the added time was not large enough to be recorded differently in the output files.

Mbgrid: Improved output to shell so that it shows the min max values from each
contributing input file whether verbose is specified or not.