Use of Local Transverse Mercator Projection for grids, visualization, and TRN
MBgrid and mbmosaic: Added an option to generate grids and mosaics in a projected
coordinate system defined as a Local Transverse Mercator (LTM) projection with the  origin
at the center of the grid or mosaic.

MBgrdviz, mbeditviz, mbnavadjust: The mbview visualization library now displays grids
defined in geographic coordinates (longitude and latitude) using a Local Transverse
Mercator projection with the origin at the center of the grid or mosaic. This replaces the
prior use of UTM projections, which introduced rotation and distortion when the grid was
located far from the center of the UTM zone.

MBtrnpp: and TRN: Changed TerrainNav, supporting libraries, and the program mbtrnpp to
calculate projections using the Proj package rather than the GCTP package.

MBtrnpp: Now works with reference topography models defined in a Local Transverse
Mercator (LTM) projection.